pyproject.toml
[project]
name = "virtual-economy-service"
version = "0.1.0"
description = "Virtual Economy Service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.2",
    "sqlalchemy>=2.0.44",
    "pydantic>=2.10.4",
    "alembic>=1.14.0",
    "dependency-injector>=4.48.2",
    "pydantic-settings>=2.12.0",
    "celery>=5.5.3",
    "pyjwt>=2.10.1",
    "uvicorn>=0.38.0",
    "redis>=4.2.0,<7",
    "asyncpg>=0.30.0",
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "slowapi>=0.1.9",
    "aiostream>=0.7.1",
]

[dependency-groups]
dev = [
    "psycopg2>=2.9.11",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.5",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]


line-length = 120
indent-width = 4

target-version = "py313"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

skip-magic-trailing-comma = false

line-ending = "auto"

docstring-code-format = false

docstring-code-line-length = "dynamic"

[tool.black]
line-length = 120
target-version = ["py313"]
include = '\.pyi?$'
